requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
requests==2.31.0
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.1.0
tensorflow==2.13.0
joblib==1.3.2
python-dotenv==1.0.0
shap==0.42.1
lime==0.2.0.1